테스트(#6117): 규칙 단위 테스트를 tests/cases 로 옮긴다
CI 의 source-side 테스트 총량 동결 게이트가 걸렸다 — `src/` 안에 단위
테스트 3건을 추가해 4221 → 4224 로 늘렸다(`[RustUnitTier] PR base 대비
source-side test 총량 증가 금지`). 저장소 정책상 새 테스트는 통합 테스트로
가야 한다.

`line_decoration_trim_target` 을 `pub` 으로 올려 통합 테스트에서 부를 수
있게 하고, 같은 3건을 `tests/cases/issue_6117_layer_decoration_trim.rs`
로 옮겼다. 판별 내용은 그대로다 — 말미 공백 트림 / 문단 끝 run 예외 /
공백뿐인 꼬리 run 무시.

검증: `rust-unit-test-tiers --check --base-ref upstream/devel` 4221 로
base 와 동일, 옮긴 3건 통과, clippy -D warnings·wasm-pack 통과.

(cherry picked from commit 13f45de98c68b83786448759006d1116da1614dd)
